requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
gunicorn==23.0.0
mysqlclient==2.2.4
PyJWT==2.9.0
pytz==2024.2
pyzmq==26.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
